test(ios): isolate #4394 caret focus repro onto an iOS 18 capability
The regression only reproduces on iOS 18's Safari touch-adjustment heuristic.
Move the test into its own spec file (caretFocus.ts) and give BrowserStack two
capabilities: the existing iOS 17 device runs the whole suite except this spec,
and a new iOS 18 device (iPhone 16 Pro Max) runs only this spec. This scopes the
intended failure and keeps the rest of the suite off iOS 18.
