+++ src/e2e/iOS/config/wdio.browserstack.conf.ts
@@ -17,6 +17,48 @@ if (!process.env.BROWSERSTACK_ACCESS_KEY) {
 const user = process.env.BROWSERSTACK_USERNAME
 const date = new Date().toISOString().slice(0, 10)
 
+// The #4394 caret focus regression only reproduces on iOS 18's Safari touch-adjustment heuristic, so
+// that single spec runs on an iOS 18 device while the rest of the suite stays on the default iOS 17.
+const caretFocusSpec = path.resolve(process.cwd(), 'src/e2e/iOS/__tests__/caretFocus.ts')
+
+const bstackOptions = {
+  projectName: process.env.BROWSERSTACK_PROJECT_NAME || 'em',
+  buildName: process.env.BROWSERSTACK_BUILD_NAME || `Local - ${user} - ${date}`,
+  local: true,
+  debug: true,
+  networkLogs: true,
+  consoleLogs: 'verbose',
+  idleTimeout: 60,
+}
+
+// Run the whole suite except the #4394 regression on the default device.
+const suiteCapability = {
+  ...baseConfig.baseCapabilities,
+  exclude: [caretFocusSpec],
+  'appium:deviceName': 'iPhone 15 Plus',
+  'appium:platformVersion': '17',
+  'bstack:options': {
+    ...bstackOptions,
+    deviceName: 'iPhone 15 Plus',
+    osVersion: '17',
+    sessionName: 'iOS Safari Tests',
+  },
+}
+
+// Run only the #4394 regression, which requires iOS 18 to reproduce.
+const caretFocusCapability = {
+  ...baseConfig.baseCapabilities,
+  specs: [caretFocusSpec],
+  'appium:deviceName': 'iPhone 16 Pro Max',
+  'appium:platformVersion': '18',
+  'bstack:options': {
+    ...bstackOptions,
+    deviceName: 'iPhone 16 Pro Max',
+    osVersion: '18',
+    sessionName: 'iOS Safari Tests (iOS 18)',
+  },
+}
+
 /**
  * WDIO configuration for BrowserStack iOS testing.
  * Uses @wdio/browserstack-service for automatic tunnel management.
@@ -34,26 +76,8 @@ export const config: WebdriverIO.Config = {
   user,
   key: process.env.BROWSERSTACK_ACCESS_KEY,
 
-  // Capabilities
-  capabilities: [
-    {
-      ...baseConfig.baseCapabilities,
-      'appium:deviceName': 'iPhone 16 Pro Max',
-      'appium:platformVersion': '18',
-      'bstack:options': {
-        deviceName: 'iPhone 16 Pro Max',
-        osVersion: '18',
-        projectName: process.env.BROWSERSTACK_PROJECT_NAME || 'em',
-        buildName: process.env.BROWSERSTACK_BUILD_NAME || `Local - ${user} - ${date}`,
-        sessionName: 'iOS Safari Tests',
-        local: true,
-        debug: true,
-        networkLogs: true,
-        consoleLogs: 'verbose',
-        idleTimeout: 60,
-      },
-    },
-  ],
+  // Capabilities (per-capability specs/exclude is a WDIO runtime feature not covered by the strict type)
+  capabilities: [suiteCapability, caretFocusCapability] as WebdriverIO.Config['capabilities'],
 
   // Services
   services: [
